Backend/Tools/requirements.txt
fastapi
pydantic
uvicorn
pika
biopython
pymsaviz
mysql-connector-python
